We don't need an env_list in tox.ini
From the tox documentation: env_list is a list of environments to run by
default (when the user does not specify anything during the invocation).
In this repo we specify which environments to run in GitHub Actions. A
blank ``tox run`` will run the ``py`` (default) test environment, as
desired.